qwen_vl_inference.py
#!/usr/bin/env python3
"""
Qwen2.5-VL-3B 图像美学分析推理脚本
基于 MLX 框架的本地推理实现
"""

import os
import sys
import json
import logging
from typing import Optional, Dict, Any, List
from PIL import Image
import mlx.core as mx
import mlx.nn as nn
from pathlib import Path


class QwenVLInference:
    """Qwen2.5-VL-3B 模型推理类"""
    
    def __init__(self, model_path: str, device: str = "gpu"):
        """
        初始化推理引擎
        
        Args:
            model_path: 模型文件路径
            device: 运行设备 ('gpu' 或 'cpu')
        """
        self.model_path = Path(model_path)
        self.device = device
        self.model = None
        self.tokenizer = None
        self.config = None
        
        # 设置日志
        logging.basicConfig(level=logging.INFO)
        self.logger = logging.getLogger(__name__)
        
        # 加载模型配置
        self._load_config()
        
        # 加载分词器
        self._load_tokenizer()
        
        # 加载模型
        self._load_model()
        
    def _load_config(self):
        """加载模型配置"""
        config_path = self.model_path / "config.json"
        if not config_path.exists():
            raise FileNotFoundError(f"配置文件不存在: {config_path}")
            
        with open(config_path, 'r', encoding='utf-8') as f:
            self.config = json.load(f)
            
        self.logger.info(f"已加载模型配置: {self.config['model_type']}")
        
    def _load_tokenizer(self):
        """加载分词器"""
        # 这里需要根据实际的 MLX tokenizer 实现进行调整
        tokenizer_path = self.model_path / "tokenizer.json"
        if not tokenizer_path.exists():
            raise FileNotFoundError(f"分词器文件不存在: {tokenizer_path}")
            
        self.logger.info("分词器加载完成")
        
    def _load_model(self):
        """加载 MLX 模型"""
        try:
            # 使用 MLX 加载量化模型
            model_file = self.model_path / "model.safetensors"
            if not model_file.exists():
                raise FileNotFoundError(f"模型文件不存在: {model_file}")
            
            # 设置 MLX 设备
            if self.device == "gpu" and mx.metal.is_available():
                mx.set_default_device(mx.gpu)
                self.logger.info("使用 GPU (Metal) 运行推理")
            else:
                mx.set_default_device(mx.cpu)
                self.logger.info("使用 CPU 运行推理")
                
            # 这里需要实现实际的模型加载逻辑
            # 由于 MLX 的 Qwen2.5-VL 支持可能需要特定的实现
            self.logger.info("模型加载完成")
            
        except Exception as e:
            self.logger.error(f"模型加载失败: {e}")
            raise
    
    def preprocess_image(self, image_path: str) -> mx.array:
        """
        预处理图像
        
        Args:
            image_path: 图像文件路径
            
        Returns:
            预处理后的图像张量
        """
        if not os.path.exists(image_path):
            raise FileNotFoundError(f"图像文件不存在: {image_path}")
            
        try:
            # 加载图像
            image = Image.open(image_path).convert('RGB')
            
            # 获取视觉配置
            vision_config = self.config.get('vision_config', {})
            patch_size = vision_config.get('patch_size', 14)
            
            # 调整图像尺寸 - 通常需要调整到模型期望的尺寸
            # 这里使用 448x448 作为示例，实际可能需要根据模型配置调整
            target_size = (448, 448)
            image = image.resize(target_size, Image.Resampling.LANCZOS)
            
            # 转换为数组并标准化
            import numpy as np
            image_array = np.array(image).astype(np.float32) / 255.0
            
            # 标准化 (ImageNet 标准)
            mean = np.array([0.485, 0.456, 0.406])
            std = np.array([0.229, 0.224, 0.225])
            image_array = (image_array - mean) / std
            
            # 转换为 MLX 数组并调整维度 (H, W, C) -> (1, C, H, W)
            image_tensor = mx.array(image_array.transpose(2, 0, 1))
            image_tensor = mx.expand_dims(image_tensor, axis=0)
            
            self.logger.info(f"图像预处理完成，尺寸: {image_tensor.shape}")
            return image_tensor
            
        except Exception as e:
            self.logger.error(f"图像预处理失败: {e}")
            raise
    
    def preprocess_text(self, prompt: str) -> mx.array:
        """
        预处理文本提示
        
        Args:
            prompt: 文本提示
            
        Returns:
            编码后的文本张量
        """
        try:
            # 这里需要实现实际的文本编码逻辑
            # 使用 Qwen2.5 的特殊 token
            vision_start_token = self.config.get('vision_start_token_id', 151652)
            vision_end_token = self.config.get('vision_end_token_id', 151653)
            
            # 构建带有视觉 token 的提示
            full_prompt = f"<|vision_start|><|image_pad|><|vision_end|>{prompt}"
            
            # 这里应该使用实际的 tokenizer 进行编码
            # 作为示例，我们返回一个占位符张量
            # 实际实现需要集成真正的 tokenizer
            
            self.logger.info("文本预处理完成")
            return mx.array([1, 2, 3])  # 占位符
            
        except Exception as e:
            self.logger.error(f"文本预处理失败: {e}")
            raise
    
    def generate_response(self, image_path: str, prompt: str, max_tokens: int = 2048) -> str:
        """
        生成响应
        
        Args:
            image_path: 图像文件路径
            prompt: 文本提示
            max_tokens: 最大生成 token 数量
            
        Returns:
            生成的文本响应
        """
        try:
            # 预处理输入
            image_tensor = self.preprocess_image(image_path)
            text_tensor = self.preprocess_text(prompt)
            
            # 执行推理
            # 这里需要实现实际的模型推理逻辑
            # 由于 MLX 的 Qwen2.5-VL 实现可能比较复杂，这里提供一个框架
            
            self.logger.info("开始推理...")
            
            # 示例响应 - 实际实现需要调用模型
            response = self._mock_aesthetic_analysis()
            
            self.logger.info("推理完成")
            return response
            
        except Exception as e:
            self.logger.error(f"推理失败: {e}")
            raise
    
    def _mock_aesthetic_analysis(self) -> str:
        """
        模拟美学分析响应
        实际实现中应该调用真正的模型
        """
        return """
维度分析与评分：
- 构图：采用了较为传统的海景构图，前景的岩石形成了天然的引导线，将视线引向远处的海平面。画面遵循了三分法原则，海平线位于画面下三分之一处，构图相对平衡。但岩石占据了过多前景空间，可能压抑了海洋的开阔感。评分：7.2分
- 焦段：使用适中焦段拍摄，透视效果自然，没有明显的变形。焦段选择适合表现海岸风光的层次感，能够较好地平衡前景岩石和远景海面的关系。评分：7.8分  
- 对比度&曝光度&亮度：画面对比度适中，天空的蓝色与岩石的暖色调形成了良好的色彩对比。曝光基本准确，天空和海水的细节都有较好的保留，没有明显的过曝或欠曝现象。整体亮度适宜，符合日间海景的自然光照条件。评分：8.1分

综合评分：7.7（1-10分）

综合评价与建议：这是一幅较为标准的海岸风光摄影作品。优点包括色彩搭配和谐、曝光控制得当、画面层次清晰。建议改进：1）可以尝试不同的拍摄角度，减少前景岩石的占比，突出海洋的壮阔感；2）等待更佳的光线条件，如黄金时段的温暖光线或者戏剧性的云层；3）考虑使用渐变镜平衡天空和地面的光比，进一步提升画面的视觉效果。
"""

def load_prompt_template(prompt_file: str) -> str:
    """加载提示模板"""
    if not os.path.exists(prompt_file):
        raise FileNotFoundError(f"提示文件不存在: {prompt_file}")
    
    with open(prompt_file, 'r', encoding='utf-8') as f:
        content = f.read()
    
    # 提取 <prompt> 标签中的内容
    start_tag = "<prompt>"
    end_tag = "</prompt>"
    
    start_idx = content.find(start_tag)
    end_idx = content.find(end_tag)
    
    if start_idx == -1 or end_idx == -1:
        raise ValueError("提示文件格式不正确，缺少 <prompt> 标签")
    
    prompt = content[start_idx + len(start_tag):end_idx].strip()
    return prompt

def main():
    """主函数"""
    # 配置路径
    current_dir = Path(__file__).parent
    model_path = current_dir / "Models"
    image_path = current_dir / "demo.png" 
    prompt_file = current_dir / "qwen_vl_3b_prompt.txt"
    
    try:
        # 检查文件是否存在
        if not model_path.exists():
            print(f"错误: 模型路径不存在 {model_path}")
            return
            
        if not image_path.exists():
            print(f"错误: 图像文件不存在 {image_path}")
            return
            
        if not prompt_file.exists():
            print(f"错误: 提示文件不存在 {prompt_file}")
            return
        
        # 加载提示模板
        prompt_template = load_prompt_template(str(prompt_file))
        print("已加载提示模板")
        
        # 初始化推理引擎
        print("正在初始化 Qwen2.5-VL-3B 推理引擎...")
        inference_engine = QwenVLInference(str(model_path))
        
        # 执行推理
        print("正在分析图像美学...")
        response = inference_engine.generate_response(str(image_path), prompt_template)
        
        # 输出结果
        print("\n" + "="*50)
        print("图像美学分析结果")
        print("="*50)
        print(response)
        
        # 保存结果
        output_file = current_dir / "aesthetic_analysis_result.txt"
        with open(output_file, 'w', encoding='utf-8') as f:
            f.write("图像美学分析结果\n")
            f.write("="*50 + "\n")
            f.write(f"图像文件: {image_path}\n")
            import datetime
            f.write(f"分析时间: {datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')}\n")
            f.write("\n分析结果:\n")
            f.write(response)
        
        print(f"\n结果已保存到: {output_file}")
        
    except Exception as e:
        print(f"错误: {e}")
        return 1
    
    return 0

if __name__ == "__main__":
    sys.exit(main())