requirements.txt
mlx>=0.19.0
mlx-lm>=0.19.0
pillow>=10.0.0
numpy>=1.24.0
safetensors>=0.4.0
transformers>=4.41.0
sentencepiece>=0.2.0